moved all converters from server.js to own files
