+++ lib/server.js
@@ -1,20 +1,29 @@
-const fs             = require('fs');
-const crypto         = require('crypto');
-const WebSocket      = require('ws');
-const bodyParser     = require('body-parser');
-const SERVICES       = require('./services');
-const PANELS         = require('./panels');
-const multer         = require('multer');
-const mime           = require('mime');
-const yaml           = require('js-yaml');
-const request        = require('request');
-const jstz           = require('jstimezonedetect');
-const utils          = require('./converters/utils');
-const processBlind   = require('./converters/cover').processBlind;
-const processSocket  = require('./converters/switch').processSocket;
-const converterLight = require('./converters/light');
+const fs                   = require('fs');
+const crypto               = require('crypto');
+const WebSocket            = require('ws');
+const bodyParser           = require('body-parser');
+const SERVICES             = require('./services');
+const PANELS               = require('./panels');
+const multer               = require('multer');
+const mime                 = require('mime');
+const yaml                 = require('js-yaml');
+const request              = require('request');
+const jstz                 = require('jstimezonedetect');
+const utils                = require('./converters/utils');
+const processBlind         = require('./converters/cover').processBlind;
+const processSocket        = require('./converters/switch').processSocket;
+const converterLight       = require('./converters/light');
+const processBinarySensors = require('./converters/binary_sensor');
+const processSensors       = require('./converters/sensor');
+const processMediaPlayer   = require('./converters/media_player').processMediaPlayer;
+const processThermostat    = require('./converters/climate').processThermostat;
+const converterWeather     = require('./converters/weather');
+const processImage         = require('./converters/camera').processImage;
+const processLock          = require('./converters/lock').processLock;
+const processLocation      = require('./converters/geo_location').processLocation;
 
-const bindings       = require('./bindings');
+const bindings   = require('./bindings');
+
 
 const TIMEOUT_PASSWORD_ENTER = 180000; // 3 min
 const TIMEOUT_AUTH_CODE = 10000; // 10sec
@@ -134,22 +143,22 @@ class WebServer {
             [Types.hue]:                    converterLight.processLightAdvanced.bind(this),
             [Types.rgb]:                    converterLight.processLightAdvanced.bind(this),
             [Types.rgbSingle]:              converterLight.processLightAdvanced.bind(this),
-            [Types.motion]:                 this._processMotion.bind(this),
-            [Types.window]:                 this._processWindow.bind(this),
-            [Types.windowTilt]:             this._processWindowTilt.bind(this),
-            [Types.door]:                   this._processDoor.bind(this),
+            [Types.motion]:                 processBinarySensors.processMotion.bind(this),
+            [Types.window]:                 processBinarySensors.processWindow.bind(this),
+            [Types.windowTilt]:             processSensors.processWindowTilt.bind(this),
+            [Types.door]:                   processBinarySensors.processDoor.bind(this),
             [Types.button]:                 processSocket.bind(this),
-            [Types.temperature]:            this._processTemperature.bind(this),
-            [Types.lock]:                   this._processLock.bind(this),
-            [Types.thermostat]:             this._processThermostat.bind(this),
+            [Types.temperature]:            processSensors.processTemperature.bind(this),
+            [Types.lock]:                   processLock.bind(this),
+            [Types.thermostat]:             processThermostat.bind(this),
             [Types.blind]:                  processBlind.bind(this),
             [Types.blindButtons]:           processBlind.bind(this),
-            [Types.weatherForecast]:        this._processWeather.bind(this),
-            [Types.accuWeatherForecast]:    this._processAccuWeather.bind(this),
-            [Types.location]:               this._processLocation.bind(this),
-            [Types.location_one]:           this._processLocation.bind(this),
-            [Types.media]:                  this._processMediaPlayer.bind(this),
-            [Types.image]:                  this._processImage.bind(this),
+            [Types.weatherForecast]:        converterWeather.processWeather.bind(this),
+            [Types.accuWeatherForecast]:    converterWeather.processAccuWeather.bind(this),
+            [Types.location]:               processLocation.bind(this),
+            [Types.location_one]:           processLocation.bind(this),
+            [Types.media]:                  processMediaPlayer.bind(this),
+            [Types.image]:                  processImage.bind(this),
         };
 
 
@@ -324,1086 +333,6 @@ class WebServer {
     }
 
     // ------------------------------- START OF CONVERTERS ---------------------------------------- //
-
-    _processWeather(id, control, name, room, func, _obj) {
-        const entity = this._processCommon(name, room, func, _obj, 'weather');
-
-        // - weather => STATE any-text(no icon)/clear-night/cloudy/fog/hail/lightning/lightning-rainy/partlycloudy/pouring/rainy/snowy/snowy-rainy/sunny/windy/windy-variant, attributes: [temperature, pressure, humidity, wind_speed, wind_bearing, forecast]
-        //        forecast is an array with max 5 items [{datetime: something for new Date(aa), temperature, templow, condition(see STATE), precipitation}, {...}]
-
-        let state = control.states.find(s => s.id && s.name === 'ICON');
-        entity.context.STATE = {getId: null};
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-        entity.context.ATTRIBUTES = [];
-
-        state = control.states.find(s => s.id && s.name === 'TEMP');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'temperature', getId: state.id});
-            this._addID2entity(state.id, entity);
-        } else {
-            state = control.states.find(s => s.id && s.name === 'TEMP_MAX');
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'temperature', getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'PRESSURE');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'pressure', getId: state.id});
-            this._addID2entity(state.id, entity);
-        } else {
-            state = control.states.find(s => s.id && s.name === 'PRESSURE0');
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'pressure', getId: state.id});
-                this._addID2entity(state.id, entity);
-                state.id = null; // do not detect it later
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'HUMIDITY');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'humidity', getId: state.id});
-            this._addID2entity(state.id, entity);
-        } else {
-            state = control.states.find(s => s.id && s.name === 'HUMIDITY0');
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'humidity', getId: state.id});
-                this._addID2entity(state.id, entity);
-                state.id = null; // do not detect it later
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'WIND_SPEED');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'wind_speed', getId: state.id});
-            this._addID2entity(state.id, entity);
-        } else {
-            state = control.states.find(s => s.id && s.name === 'WIND_SPEED0');
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'wind_speed', getId: state.id});
-                this._addID2entity(state.id, entity);
-                state.id = null; // do not detect it later
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'WIND_DIRECTION'); // in °
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'wind_bearing', getId: state.id});
-            this._addID2entity(state.id, entity);
-        } else {
-            state = control.states.find(s => s.id && s.name === 'WIND_DIRECTION0'); // in °
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'wind_bearing', getId: state.id});
-                this._addID2entity(state.id, entity);
-                state.id = null; // do not detect it later
-            }
-        }
-
-        // forecast {datetime: something for new Date(aa), temperature, templow, condition(see STATE), precipitation}
-        // day 0
-        let somethingFound;
-        let hassCounter = -1;
-
-        for (let day = 0; day < 6; day++) {
-            let dayShiftId;
-            somethingFound = false;
-            state = control.states.find(s => s.id && s.name === 'ICON' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.condition`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-            state = control.states.find(s => s.id && s.name === 'TEMP' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.temperature`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            } else {
-                state = control.states.find(s => s.id && s.name === 'TEMP_MAX' + day);
-                if (state && state.id) {
-                    if (!somethingFound) {
-                        hassCounter++;
-                        somethingFound = true;
-                    }
-                    dayShiftId = dayShiftId || state.id;
-                    entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.temperature`, getId: state.id});
-                    this._addID2entity(state.id, entity);
-                }
-            }
-
-            state = control.states.find(s => s.id && s.name === 'TEMP_MIN' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.templow`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-            state = control.states.find(s => s.id && s.name === 'PRECIPITATION_CHANCE' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.precipitation_probability`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-            state = control.states.find(s => s.id && s.name === 'PRECIPITATION' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.precipitation`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-            if (somethingFound) {
-                state = control.states.find(s => s.id && s.name === 'DATE' + day);
-                if (state && state.id) {
-                    entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.datetime`, getId: state.id});
-                    this._addID2entity(state.id, entity);
-                } else if (dayShiftId) {
-                    entity.context.ATTRIBUTES.push({
-                        attribute: `forecast.${hassCounter}.datetime`,
-                        dayShift: hassCounter,
-                        getId: dayShiftId,
-                        getParser: (entity, attr) => {
-                            const date = new Date();
-                            attr.dayShift && date.setDate(date.getDate() + attr.dayShift);
-                            this.setJsonAttribute(entity.attributes, attr.attribute, date.toISOString());
-                        }
-                    });
-                }
-            } else if (hassCounter >= 0) {
-                break;
-            }
-        }
-
-        return [entity];
-    }
-
-    _processAccuWeather(id, control, name, room, func, _obj) {
-        const entity = this._processCommon(name, room, func, _obj, 'weather');
-
-        // - weather => STATE any-text(no icon)/clear-night/cloudy/fog/hail/lightning/lightning-rainy/partlycloudy/pouring/rainy/snowy/snowy-rainy/sunny/windy/windy-variant, attributes: [temperature, pressure, humidity, wind_speed, wind_bearing, forecast]
-        //        forecast is an array with max 5 items [{datetime: something for new Date(aa), temperature, templow, condition(see STATE), precipitation}, {...}]
-
-        let state = control.states.find(s => s.id && s.name === 'STATE');
-        entity.context.STATE = {getId: null};
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-        entity.context.ATTRIBUTES = [];
-
-        state = control.states.find(s => s.id && s.name === 'TEMP');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'temperature', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'ICON_URL');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'icon_url', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'TIME_SUNRISE');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'sunrise', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'TIME_SUNSET');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'sunset', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'PRESSURE');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'pressure', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'HUMIDITY');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'humidity', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'WIND_SPEED');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'wind_speed', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'WIND_DIRECTION_STR'); // compass
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'wind_bearing', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        // forecast {datetime: something for new Date(aa), temperature, templow, condition(see STATE), precipitation}
-        // day 0
-        let somethingFound;
-        let hassCounter = -1;
-
-        for (let day = 0; day < 6; day++) {
-            let dayShiftId;
-            somethingFound = false;
-            state = control.states.find(s => s.id && s.name === 'STATE' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.condition`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-            state = control.states.find(s => s.id && s.name === 'ICON_URL' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.icon_url`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-
-            state = control.states.find(s => s.id && s.name === 'TEMP_MAX' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.temperature`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-
-            state = control.states.find(s => s.id && s.name === 'TEMP_MIN' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.templow`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-
-            state = control.states.find(s => s.id && s.name === 'PRECIPITATION' + day);
-            if (state && state.id) {
-                if (!somethingFound) {
-                    hassCounter++;
-                    somethingFound = true;
-                }
-                dayShiftId = dayShiftId || state.id;
-                entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.precipitation`, getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-
-            if (somethingFound) {
-                state = control.states.find(s => s.id && s.name === 'DATE' + day);
-                if (state && state.id) {
-                    entity.context.ATTRIBUTES.push({attribute: `forecast.${hassCounter}.datetime`, getId: state.id});
-                    this._addID2entity(state.id, entity);
-                } else if (dayShiftId) {
-                    entity.context.ATTRIBUTES.push({
-                        attribute: `forecast.${hassCounter}.datetime`,
-                        dayShift: hassCounter,
-                        getId: dayShiftId,
-                        getParser: (entity, attr) => {
-                            const date = new Date();
-                            attr.dayShift && date.setDate(date.getDate() + attr.dayShift);
-                            this.setJsonAttribute(entity.attributes, attr.attribute, date.toISOString());
-                        }
-                    });
-                }
-            } else if (hassCounter >= 0) {
-                break;
-            }
-        }
-
-        return [entity];
-    }
-
-    _processLocation(id, control, name, room, func, _obj) {
-        // - geo_location => attributes: [latitude, longitude, passive, icon, radius, entity_picture, gps_accuracy, source]
-        const entity = this._processCommon(name, room, func, _obj, 'geo_location');
-
-        entity.context.STATE = {getId: null};
-        entity.context.ATTRIBUTES = [];
-        entity.attributes.icon = 'mdi:crosshairs-gps';
-
-        let state = control.states.find(s => s.id && s.name === 'GPS');
-        if (state && state.id) {
-            const getParser = function (entity, attr, state) {
-                if (!state) {
-                    state = {val: ''};
-                }
-                const value = state.val;
-                let parts = value.split(';');
-                if (parts.length !== 2) {
-                    parts = value.split(',');
-                }
-                if (parts.length === 2) {
-                    entity.attributes.latitude = parseFloat(parts[0]);
-                    entity.attributes.longitude = parseFloat(parts[1]);
-                }
-            };
-
-            entity.context.ATTRIBUTES.push({
-                attribute: 'latitude',
-                getId: state.id,
-                getParser: getParser,
-                historyParser: (id, val) => {
-                    val = (val || '').toString();
-                    let parts = val.toString().split(';');
-                    if (parts.length !== 2) {
-                        parts = val.toString().split(',');
-                    }
-                    return parts[0];
-                }
-            }, {
-                attribute: 'longitude',
-                getId: state.id,
-                getParser: getParser,
-                historyParser: (id, val) => {
-                    val = (val || '').toString();
-                    let parts = val.toString().split(';');
-                    if (parts.length !== 2) {
-                        parts = val.toString().split(',');
-                    }
-                    return parts[1];
-                }
-            });
-            this._addID2entity(state.id, entity);
-        } else {
-            state = control.states.find(s => s.id && s.name === 'LATITUDE');
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'latitude', getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-
-            state = control.states.find(s => s.id && s.name === 'LONGITUDE');
-            if (state && state.id) {
-                entity.context.ATTRIBUTES.push({attribute: 'longitude', getId: state.id});
-                this._addID2entity(state.id, entity);
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'RADIUS');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'radius', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'ACCURACY');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'gps_accuracy', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        entity.attributes.source = 'gps';
-
-        return [entity];
-    }
-
-    _processImage(id, control, name, room, func, _obj) {
-        // - geo_location => attributes: [latitude, longitude, passive, icon, radius, entity_picture, gps_accuracy, source]
-        const entity = this._processCommon(name, room, func, _obj, 'camera');
-
-        entity.context.STATE = {getId: null};
-        entity.context.ATTRIBUTES = [];
-        entity.attributes.icon = 'mdi:camera-outline';
-
-        const state = control.states.find(s => s.id && s.name === 'URL');
-        if (state && state.id) {
-            entity.context.STATE = {getValue: 'on'};
-            entity.context.ATTRIBUTES = [{getId: state.id, attribute: 'url'}];
-            entity.attributes.code_format = 'number';
-
-            entity.attributes.access_token = crypto
-                .createHmac('sha256', (Math.random() * 1000000000).toString())
-                .update(Date.now().toString())
-                .digest('hex');
-
-            entity.attributes.model_name = 'Simulated URL';
-            entity.attributes.brand = 'ioBroker';
-            entity.attributes.motion_detection = false;
-            this._addID2entity(state.id, entity);
-        }
-
-        return [entity];
-    }
-
-    _processMediaPlayer(id, control, name, room, func, _obj, objects) {
-        // - media_player =>
-        // STATE on/off/playing/paused/idle/standby/unknown,
-        //     attributes: [
-        //          media_content_type(music/game/music/tvshow/...), // not supported
-        //          entity_picture(as cover),
-        //          supported_features,
-        //          is_volume_muted,
-        //          volume_level,
-        //          media_duration,
-        //          media_position,
-        //          media_position_updated_at,
-        //          media_title,
-        //          media_artist,
-        //          media_series_title, // not supported
-        //          media_season,
-        //          media_episode,
-        //          app_name, // not supported
-        //          source, // not supported
-        //          source_list, // not supported
-        //          sound_mode, // not supported
-        //          sound_mode_list // not supported
-        //          ],
-        //     commands:
-        //          media_play_pause,
-        //          media_stop,
-        //          media_next_track,
-        //          media_previous_track,
-        //          volume_set(volume_level),
-        //          turn_off,
-        //          turn_on,
-        //          volume_down,
-        //          volume_mute(is_volume_muted),
-        //          volume_up,
-        //          select_source(source), // not supported
-        //          select_sound_mode(sound_mode), // not supported
-        //     features for supported_features:
-        //          PAUSE 0x1,
-        //          SEEK 0x2,
-        //          volume_set 0x4,
-        //          volume_mute 0x8,
-        //          media_previous_track 0x10,
-        //          media_next_track 0x20,
-        //          turn_on 0x80,
-        //          turn_off 0x100,
-        //          play_media 0x200,
-        //          volume_down/volume_up 0x400,
-        //          select_source 0x800,
-        //          stop 0x1000,
-        //          clear_playlist 0x2000,
-        //          play (0x4000),
-        //          shuffle 0x8000,
-        //          select_sound_mode (0x10000)
-
-        const entity = this._processCommon(name, room, func, _obj, 'media_player');
-
-        entity.context.STATE = {getId: null};
-        entity.context.ATTRIBUTES = [];
-        entity.context.COMMANDS = [];
-        entity.attributes.icon = 'mdi:play-network';
-        entity.attributes.media_content_type = 'unknown';
-        entity.attributes.supported_features = 0;
-        let playPauseCommand;
-
-        const parsePlayPauseCommand = (entity, command, params, user) => {
-            let id = command.setId;
-            let val = true;
-            if (entity.state === 'playing') {
-                //ok, need to pause.
-                if (command.pauseId) {
-                    id = command.pauseId;
-                } else {
-                    val = false;
-                }
-                entity.state = 'paused';
-            } else {
-                entity.state = 'playing';
-            }
-            return new Promise((resolve, reject) =>
-                this.adapter.setForeignState(id, val, false, {user}, err =>
-                    err ? reject(err) : resolve()));
-        };
-
-        let state = control.states.find(s => s.id && s.name === 'STATE');
-        if (state && state.id) {
-            entity.context.STATE.setId = state.id;
-            entity.context.STATE.getId = state.id;
-            entity.context.STATE.getParser = (entity, attr, state) => {
-                state = state || {val: null};
-                // on/off/playing/paused/idle/standby/unknown,
-                if (state.val === true || state.val === 1 || state.val === 'play') {
-                    entity.state = 'playing';
-                } else if (state.val === false || state.val === 0 || state.val === 'pause') {
-                    entity.state = 'paused';
-                } else if (state.val === 'stop' || state.val === 2) {
-                    entity.state = 'idle';
-                }
-            };
-            entity.context.STATE.historyParser = (id, val) => {
-                if (val === true || val === 1 || val === 'play') {
-                    return 'playing';
-                }
-                if (val === false || val === 0 || val === 'pause') {
-                    return 'paused';
-                }
-                return 'idle';
-            };
-
-            if (objects[state.id] && objects[state.id].common && objects[state.id].common.write) {
-                playPauseCommand = {
-                    service: 'media_play_pause',
-                    setId: state.id,
-                    parseCommand: parsePlayPauseCommand
-                };
-                entity.context.COMMANDS.push(playPauseCommand);
-                entity.attributes.supported_features |= 0x4000;
-            }
-            this._addID2entity(state.id, entity);
-        }
-
-        if (!playPauseCommand) {
-            state = control.states.find(s => s.id && s.name === 'PLAY');
-            if (state && state.id) {
-                playPauseCommand = {
-                    service: 'media_play_pause',
-                    setId: state.id,
-                    parseCommand: parsePlayPauseCommand
-                };
-                entity.context.COMMANDS.push(playPauseCommand);
-                entity.attributes.supported_features |= 0x4000;
-            }
-
-            state = control.states.find(s => s.id && s.name === 'PAUSE');
-            if (!state || !state.id) { //try to use STOP for Pause.
-                state = control.states.find(s => s.id && s.name === 'STOP');
-            }
-            if (state && state.id) {
-                entity.attributes.supported_features |= 0x1;
-                if (playPauseCommand) {
-                    playPauseCommand.pauseId = state.id;
-                } else {
-                    entity.context.COMMANDS.push({
-                        service: 'media_play_pause',
-                        pauseId: state.id,
-                        parseCommand: parsePlayPauseCommand
-                    });
-                }
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'STOP');
-        if (state && state.id) {
-            entity.context.COMMANDS.push({
-                service: 'media_stop',
-                stopId: state.id,
-                parseCommand: (entity, command, params, user) => new Promise((resolve, reject) =>
-                    this.adapter.setForeignState(command.stopId, true, false, {user}, err =>
-                        err ? reject(err) : resolve()))
-            });
-            entity.attributes.supported_features |= 0x1000;
-        }
-
-        state = control.states.find(s => s.id && s.name === 'SHUFFLE');
-        if (state && state.id) {
-            entity.context.COMMANDS.push({
-                service: 'shuffle_set',
-                setId: state.id,
-                parseCommand: (entity, command, params, user) => new Promise((resolve, reject) =>
-                    this.adapter.setForeignState(command.setId, true, false, {user}, err =>
-                        err ? reject(err) : resolve()))
-            });
-            entity.context.ATTRIBUTES.push({
-                attribute: 'shuffle',
-                getId: state.id,
-            });
-            this._addID2entity(state.id, entity);
-            entity.attributes.supported_features |= 0x8000;
-        }
-
-        state = control.states.find(s => s.id && s.name === 'NEXT');
-        if (state && state.id) {
-            entity.context.COMMANDS.push({
-                service: 'media_next_track',
-                setId: state.id,
-                parseCommand: (entity, command, params, user) =>
-                    new Promise((resolve, reject) =>
-                        this.adapter.setForeignState(command.setId, true, false, {user}, err =>
-                            err ? reject(err) : resolve()))
-
-            });
-            entity.attributes.supported_features |= 0x20;
-        }
-
-        state = control.states.find(s => s.id && s.name === 'PREV');
-        if (state && state.id) {
-            entity.context.COMMANDS.push({
-                service: 'media_previous_track',
-                setId: state.id,
-                parseCommand: (entity, command, params, user) =>
-                    new Promise((resolve, reject) =>
-                        this.adapter.setForeignState(command.setId, true, false, {user}, err =>
-                            err ? reject(err) : resolve()))
-
-            });
-            entity.attributes.supported_features |= 0x10;
-        }
-        let getVolumeId;
-        state = control.states.find(s => s.id && s.name === 'VOLUME_ACTUAL');
-        if (state && state.id) {
-            getVolumeId = state.id;
-        }
-        state = control.states.find(s => s.id && s.name === 'VOLUME');
-
-        let setVolumeId;
-        if (state && state.id) {
-            setVolumeId = state.id;
-            getVolumeId = getVolumeId || state.id;
-
-            entity.attributes.supported_features |= 0x4;
-            entity.attributes.supported_features |= 0x400;
-
-            const common = objects[setVolumeId].common || {};
-
-            const min = common.min !== undefined ? common.min : 0;
-            const max = common.max !== undefined ? common.max : 100;
-            const step = common.step !== undefined ? common.step : (max - min) / 10;
-
-            entity.context.COMMANDS.push({
-                service: 'volume_down',
-                setId: setVolumeId,
-                getId: getVolumeId,
-                min,
-                max,
-                step,
-                parseCommand: (entity, command, params, user) => new Promise((resolve, reject) => {
-                    // get status
-                    this.adapter.getForeignState(command.getId, (err, state) => {
-                        let val = (state && state.val) || 0;
-                        val -= command.step;
-                        if (val < command.min) {
-                            val = command.min;
-                        }
-                        this.adapter.setForeignState(command.setId, val, false, {user}, err =>
-                            err ? reject(err) : resolve());
-                    });
-                })
-            });
-
-            entity.context.COMMANDS.push({
-                service: 'volume_up',
-                setId: setVolumeId,
-                getId: getVolumeId,
-                min,
-                max,
-                step,
-                parseCommand: (entity, command, params, user) => new Promise((resolve, reject) => {
-                    // get status
-                    this.adapter.getForeignState(command.getId, (err, state) => {
-                        let val = (state && state.val) || 0;
-                        val += command.step;
-                        if (val > command.max) {
-                            val = command.max;
-                        }
-                        this.adapter.setForeignState(command.setId, val, false, {user}, err =>
-                            err ? reject(err) : resolve());
-                    });
-                })
-            });
-
-            entity.context.COMMANDS.push({
-                service: 'volume_set',
-                setId: setVolumeId,
-                min: min,
-                max: max,
-                parseCommand: (entity, command, data, user) =>
-                    new Promise((resolve, reject) => {
-                        let volume = data.service_data.value !== undefined ? data.service_data.value : data.service_data.volume_level;
-                        volume = volume * command.max + command.min; //scale from [0.0,1.0] to [min,max]
-                        this.adapter.setForeignState(command.setId, volume, false, {user}, err =>
-                            err ? reject(err) : resolve());
-                    })
-            });
-
-            entity.context.ATTRIBUTES.push({
-                attribute: 'volume_level',
-                getId: state.id,
-                min: min,
-                max: max,
-                getParser: function (entity, attr, state) {
-                    state = state || {val: null};
-                    if (state.val !== null) {
-                        entity.attributes.volume_level = (state.val - attr.min) / attr.max;
-                    }
-                }
-            });
-            entity.attributes.volume_level = 0;
-            this._addID2entity(state.id, entity);
-        } else if (getVolumeId) {
-            const common = objects[getVolumeId].common || {};
-            const min = common.min !== undefined ? common.min : 0;
-            const max = common.max !== undefined ? common.max : 100;
-            entity.context.ATTRIBUTES.push({
-                attribute: 'volume_level',
-                getId: getVolumeId,
-                min: min,
-                max: max,
-                getParser: function (entity, attr, state) {
-                    state = state || {val: null};
-                    if (state.val !== null) {
-                        entity.attributes.volume_level = (state.val - attr.min) / attr.max;
-                    }
-                }
-            });
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'MUTE');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({
-                attribute: 'is_volume_muted',
-                getId: state.id
-            });
-
-            entity.context.COMMANDS.push({
-                service: 'volume_mute',
-                setId: state.id,
-                parseCommand: (entity, command, data, user) =>
-                    new Promise((resolve, reject) =>
-                        this.adapter.setForeignState(command.setId, data.service_data.is_volume_muted, false, {user}, err =>
-                            err ? reject(err) : resolve()))
-            });
-
-            this._addID2entity(state.id, entity);
-            entity.attributes.supported_features |= 0x8;
-        }
-
-        state = control.states.find(s => s.id && s.name === 'POWER');
-        if (state && state.id) {
-            entity.context.COMMANDS.push({
-                service: 'turn_on',
-                setId: state.id,
-                parseCommand: (entity, command, params, user) =>
-                    new Promise((resolve, reject) =>
-                        this.adapter.setForeignState(command.setId, true, false, {user}, err =>
-                            err ? reject(err) : resolve()))
-            });
-
-            entity.context.COMMANDS.push({
-                service: 'turn_off',
-                setId: state.id,
-                parseCommand: (entity, command, params, user) =>
-                    new Promise((resolve, reject) =>
-                        this.adapter.setForeignState(command.setId, false, false, {user}, err =>
-                            err ? reject(err) : resolve()))
-            });
-
-            entity.context.ATTRIBUTES.push({
-                attribute: 'power',
-                getId: state.id,
-                getParser: function (entity, attr, state) {
-                    state = state || {val: null};
-                    if (state.val) {
-                        entity.attributes.power = 'on';
-                    } else {
-                        entity.attributes.power = 'off';
-                        entity.state = 'off';
-                    }
-                }
-            });
-            entity.attributes.supported_features |= 0x80;
-            entity.attributes.supported_features |= 0x100;
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'ARTIST');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'media_artist', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'TITLE');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'media_title', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'DURATION');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'media_duration', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'ELAPSED');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'media_position', getId: state.id});
-            entity.context.ATTRIBUTES.push({attribute: 'media_position_updated_at', media_position_updated_at: true});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'EPISODE');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'media_episode', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'SEASON');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'media_season', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'COVER');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES.push({attribute: 'entity_picture', getId: state.id});
-            this._addID2entity(state.id, entity);
-        }
-
-        return [entity];
-    }
-
-    _processDoor(id, control, name, room, func, _obj, objects) {
-        const entity = this._processCommon(name, room, func, _obj, 'binary_sensor');
-
-        const state = control.states.find(s => s.id && s.name === 'ACTUAL');
-        entity.context.STATE = {getId: null};
-        entity.attributes.icon = 'mdi:door';
-        entity.attributes.device_class = 'door';
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-
-        return [entity, this._processBattery(control, name, room, func, objects)];
-    }
-
-    _processWindow(id, control, name, room, func, _obj, objects) {
-        const entity = this._processCommon(name, room, func, _obj, 'binary_sensor');
-
-        const state = control.states.find(s => s.id && s.name === 'ACTUAL');
-        entity.context.STATE = {getId: null};
-        entity.attributes.icon = 'mdi:window-maximize';
-        entity.attributes.device_class = 'window';
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-
-        return [entity, this._processBattery(control, name, room, func, objects)];
-    }
-
-    _processWindowTilt(id, control, name, room, func, _obj, objects) {
-        const entity = this._processCommon(name, room, func, _obj, 'sensor');
-
-        const state = control.states.find(s => s.id && s.name === 'ACTUAL');
-        entity.context.STATE = {getId: null};
-        entity.attributes.icon = 'mdi:window-maximize';
-        entity.attributes.device_class = 'window';
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            entity.context.STATE.states = objects[state.id] && objects[state.id].common ? objects[state.id].common.states || null : null;
-            entity.context.STATE.getParser = function (entity, attr, state) {
-                state = state || {val: null};
-                entity.state = entity.context.STATE.historyParser(id, state.val);
-            };
-            entity.context.STATE.historyParser = function (id, val) {
-                let str;
-                if (entity.context.STATE.states) {
-                    str = entity.context.STATE.states[val] ? entity.context.STATE.states[val].toLowerCase() : 'error';
-                } else {
-                    //use 0 = closed, 1 = tilted, 2 = open as default from ioBroker-roles documentation, if no common.states
-                    str = val === 0 ? 'closed' : val === 1 ? 'tilted' : 'open';
-                }
-                //try to translate -> will work for open, closed, tilted.
-                return this.words[str] ? this.words[str][this.lang] || this.words[str].en : str;
-            }.bind(this);
-            this._addID2entity(state.id, entity);
-        }
-
-        return [entity, this._processBattery(control, name, room, func, objects)];
-    }
-
-    _processBattery(control, name, room, func, objects) {
-        const state = control.states.find(s => s.id && s.name === 'LOWBAT');
-        if (state && state.id) {
-            const entity = this._processCommon(name, room, func, objects[state.id], 'sensor');
-            entity.context.STATE = {getId: state.id};
-            entity.context.iobType = 'LOWBAT';
-            entity.attributes.icon = 'mdi:battery-alert';
-            entity.attributes.device_class = 'battery';
-            entity.unit_of_measurement = '%';
-            this._addID2entity(state.id, entity);
-            return entity;
-        } else {
-            return null;
-        }
-    }
-
-    _processMotion(id, control, name, room, func, _obj, objects) {
-        const entity = this._processCommon(name, room, func, _obj, 'binary_sensor');
-
-        const state = control.states.find(s => s.id && s.name === 'ACTUAL');
-        entity.context.STATE = {getId: null};
-        entity.attributes.icon = 'mdi:motion-sensor';
-        entity.attributes.device_class = 'motion';
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-
-        return [entity, this._processBattery(control, name, room, func, objects)];
-    }
-
-    _processLock(id, control, name, room, func, _obj, objects) {
-        const entity = this._processCommon(name, room, func, _obj, 'lock');
-
-        let state = control.states.find(s => s.id && s.name === 'SET');
-        entity.context.STATE = {setId: null, getId: null};
-        if (state && state.id) {
-            entity.context.STATE.setId = state.id;
-            entity.context.STATE.getId = state.id; //fallback if no getId.
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'ACTUAL');
-        if (state && state.id) {
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'OPEN');
-        if (state && state.id) {
-            entity.attributes.supported_features = 1;
-            entity.context.COMMANDS = [{
-                service: 'open',
-                setId: state.id,
-                parseCommand: (entity, command, data, user) => {
-                    return new Promise((resolve, reject) => {
-                        this.adapter.setForeignState(command.setId, true, false,
-                            {user}, err => err ? reject(err) : resolve());
-                    });
-                }
-            }];
-        }
-        return [entity, this._processBattery(control, name, room, func, objects)];
-    }
-
-    _processTemperature(id, control, name, room, func, _obj, objects) {
-        let entity;
-        let state = control.states.find(s => s.id && s.name === 'ACTUAL'); // temperature
-        if (state && state.id) {
-            entity = this._processCommon(name, room, func, objects[state.id], 'sensor');
-            entity.context.STATE = {getId: state.id};
-            entity.attributes.device_class = 'temperature';
-            this._addID2entity(state.id, entity);
-        }
-
-        state = control.states.find(s => s.id && s.name === 'SECOND'); // humidity
-        let entityHum;
-        if (state && state.id) {
-            entityHum = this._processCommon(name, room, func, objects[state.id], 'sensor');
-            entityHum.context.STATE = {getId: state.id};
-            entityHum.attributes.icon = 'mdi:mdi-water';
-            entityHum.attributes.device_class = 'humidity';
-            entityHum.attributes.unit_of_measurement = entityHum.attributes.unit_of_measurement || objects[state.id].common.unit || '%';
-            this._addID2entity(state.id, entityHum);
-        }
-
-        return [entity, entityHum, this._processBattery(control, name, room, func, objects)];
-    }
-
-    _processThermostat(id, control, name, room, func, _obj, objects) {
-        // - climate => STATE on/off, attributes: [current_temperature, operation_mode, operation_list, target_temp_step, target_temp_low, target_temp_high, min_temp, max_temp, temperature], commands:
-        const entity = this._processCommon(name, room, func, _obj, 'climate');
-
-        let state = control.states.find(s => s.id && s.name === 'POWER');
-        entity.context.STATE = {setId: null, getId: null};
-        entity.attributes.icon = 'mdi:thermostat';
-
-        if (state && state.id) {
-            entity.context.STATE.setId = state.id;
-            entity.context.STATE.getId = state.id;
-            this._addID2entity(state.id, entity);
-        }
-
-        // actual temperature
-        state = control.states.find(s => s.id && s.name === 'ACTUAL');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES = entity.context.ATTRIBUTES || [];
-            entity.context.ATTRIBUTES.push({attribute: 'current_temperature', getId: state.id});
-            this._addID2entity(state.id, entity);
-            if (objects[state.id].common && objects[state.id].common.unit) {
-                entity.attributes.unit_of_measurement = objects[state.id].common.unit;
-            }
-        }
-
-        state = control.states.find(s => s.id && s.name === 'SET');
-        if (state && state.id) {
-            entity.context.ATTRIBUTES = entity.context.ATTRIBUTES || [];
-            entity.context.ATTRIBUTES.push({attribute: 'temperature', getId: state.id});
-            entity.context.COMMANDS = [{
-                service: 'set_temperature',
-                setId: state.id,
-                parseCommand: (entity, command, data, user) =>
-                    new Promise((resolve, reject) =>
-                        this.adapter.setForeignState(command.setId, data.service_data.temperature, false, {user}, err =>
-                            err ? reject(err) : resolve()))
-            }];
-
-            if (objects[state.id].common) {
-                if (!entity.attributes.unit_of_measurement && objects[state.id].common.unit) {
-                    entity.attributes.unit_of_measurement = objects[state.id].common.unit;
-                }
-                if (objects[state.id].common.min) {
-                    entity.attributes.min_temp = objects[state.id].common.min;
-                }
-                if (objects[state.id].common.max) {
-                    entity.attributes.max_temp = objects[state.id].common.max;
-                }
-                if (objects[state.id].common.step) {
-                    entity.attributes.target_temp_step = objects[state.id].common.step;
-                }
-            }
-            this._addID2entity(state.id, entity);
-        }
-
-        // detect second entity => humidity
-        let entryHum;
-        state = control.states.find(s => s.id && s.name === 'HUMIDITY');
-        if (state && state.id) {
-            entryHum = this._processCommon(name, room, func, objects[state.id], 'sensor');
-            entryHum.context.STATE = {getId: state.id};
-            entryHum.attributes.icon = 'mdi:mdi-water';
-            entryHum.attributes.device_class = 'humidity';
-            entryHum.attributes.unit_of_measurement = entryHum.attributes.unit_of_measurement || objects[state.id].common.unit || '%';
-            this._addID2entity(state.id, entryHum);
-        }
-        return [entity, entryHum, this._processBattery(control, name, room, func, objects)];
-    }
 
     _iobState2EntityState(id, val, type) {
         type = type || this._ID2entity[id][0].context.type;
